refactor: Models general improvement
- Base model has been renamed to "BaseAPIModel"
- Module base_reponse has been renamed to _base since it will be only
  used by models and won't be imported outside it.
- _base doctring now has information about what the module includes.
- Sections in based has been divided using comments blocks.
- Docstrings from response modules has been removed, since they're
  mostly autogenerated and including docstring in each of them will be a
  lack of time.
