+++ akcli/models/__init__.py
@@ -4,12 +4,10 @@
 Pydantic models representing Akamai API responses.
 """
 
-from .base_response import BaseResponse
 from .dig_response import DigResponse
 from .translate_response import TranslateResponse
 
 __all__ = [
-    "BaseResponse",
     "DigResponse",
     "TranslateResponse",
 ]
